+++ ignite/cmd/chain_serve.go
@@ -10,7 +10,6 @@ import (
 	"github.com/ignite/cli/v28/ignite/pkg/cliui"
 	uilog "github.com/ignite/cli/v28/ignite/pkg/cliui/log"
 	cliuimodel "github.com/ignite/cli/v28/ignite/pkg/cliui/model"
-	"github.com/ignite/cli/v28/ignite/pkg/cosmosver"
 	"github.com/ignite/cli/v28/ignite/pkg/errors"
 	"github.com/ignite/cli/v28/ignite/pkg/events"
 	"github.com/ignite/cli/v28/ignite/services/chain"
@@ -76,7 +75,7 @@ production, you may want to run "appd start" manually.
 	c.Flags().BoolP(flagResetOnce, "r", false, "reset the app state once on init")
 	c.Flags().Bool(flagGenerateClients, false, "generate code for the configured clients on reset or source code change")
 	c.Flags().Bool(flagQuitOnFail, false, "quit program if the app fails to start")
-	c.Flags().StringSlice(flagBuildTags, []string{cosmosver.DefaultVersion().String()}, "parameters to build the chain binary")
+	c.Flags().StringSlice(flagBuildTags, []string{}, "parameters to build the chain binary")
 
 	return c
 }