+++ changelog.md
@@ -7,6 +7,15 @@
 - [#3839](https://github.com/ignite/cli/pull/3839) New structure for app scaffolding
 - [#3835](https://github.com/ignite/cli/pull/3835) Add `--minimal` flag to `scaffold chain` to scaffold a chain with the least amount of sdk modules
 
+### Changes
+
+- [#3899](https://github.com/ignite/cli/pull/3899) Introduce `plugin.Execute` function
+- [#3903](https://github.com/ignite/cli/pull/3903) Don't specify a default build tag and deprecate notion of app version
+
+### Bug Fixes
+
+- [#3905](https://github.com/ignite/cli/pull/3905) Fix `ignite completion`
+
 ## [`v28.1.1`](https://github.com/ignite/cli/releases/tag/v28.1.1)
 
 ### Fixes